requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
magic-filter==1.0.12
multidict==6.0.5
notion-client==2.3.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==0.21.1
pytz==2025.2
sniffio==1.3.1
typing-extensions==4.7.1
yarl==1.9.4
zipp==3.15.0